temp-sculpt-cavity-mask: Add caching
Cavity masking now builds and caches values on demand.  It
does this by using two temporary attributes:
one for the final cavity mask and another for a stroke id.
The stroke id attribute is used to detect when verts need
to recalulate their cavity mask value; its values are compared
to a global stroke id that's generated at the beginning of each
stroke.

Notes:
* Cavity mask blurring now happens with a breadth-first walk
  per-vertex.
* The "attributes" are currently just simple arrays inside
  of the Sculpt struct; later on they'll use the planned
  sculpt temporary attribute api.
* I wasn't able to find a small-buffer-optimized queue class,
  so the queue logic is all manual (it uses a blender::Vector).
* sculpt_automasking.c is now compiled with C++.